switch away from sigmoid.

+++ DECCaTNet/DECCaTNet_model/fine_tuning.py
@@ -208,7 +208,7 @@ def train_epoch(model, train_loader, device, loss_func, optimizer,disable):
         #print(f'target variables before changign them {y}')
         y = torch.Tensor([ 1 if not elem else 0 for elem in y]) # TODO: this is only works with n_classes = 2
         #print(f'target variables after changing: {y}')
-        #y = y.type(torch.LongTensor)
+        y = y.type(torch.LongTensor)
         x, y = x.to(device), y.to(device)
         optimizer.zero_grad()
         # forward pass
@@ -250,7 +250,7 @@ def validate_epoch(model, val_loader, device, loss_func,disable):
         model.eval()  # tell model it is evaluation time
         for x, y in tqdm(val_loader, disable=disable):
             y = torch.Tensor([ 1 if not elem else 0 for elem in y])
-            #y = y.type(torch.LongTensor)
+            y = y.type(torch.LongTensor)
             x, y = x.to(device), y.to(device)
             # get predictions
             pred = model(x)
